refactor for merge
